delete post security added

+++ app/Http/Controllers/PostsController.php
@@ -26,6 +26,9 @@ class PostsController extends Controller
     }
     public function destroy(Post $post)
     {
+        if (!$post->ownedBy(auth()->user())) {
+            return back()->with('success', "You can't delete other post");
+        }
         $post->delete();
         return back()->with('success', "Post deleted successfully");
     }
